requirements.txt
Flask
Flask-SQLAlchemy==3.0.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
Werkzeug
gunicorn==23.0.0
bleach==6.2.0
